Refactor personio api
