+++ app/utils/workTimeFormValidator.ts
@@ -1,11 +1,11 @@
 import type { ZodSchema, ZodTypeDef } from "zod";
 import { z } from "zod";
 import type { Time } from "./dateTimeUtils";
-import { YYYY_MM_DD_FORMAT, timeSchema } from "./dateTimeUtils";
+import { YYYY_MM_DD_FORMAT, timeSchema, timeToMinutes } from "./dateTimeUtils";
 
 export type WorkTimeFormData = {
   startTime: Time;
-  breakTime: Time;
+  breakTime: number;
   endTime: Time;
   date: string;
   _action: "POST" | "PATCH" | "DELETE";
@@ -13,7 +13,7 @@ export type WorkTimeFormData = {
 
 export const workTimeFormDataSchema = z.object({
   startTime: timeSchema,
-  breakTime: timeSchema,
+  breakTime: timeSchema.transform((time) => timeToMinutes(time)),
   endTime: timeSchema,
   date: z.string().regex(YYYY_MM_DD_FORMAT),
   _action: z.enum(["POST", "PATCH", "DELETE"]),
